+++ src/api.ts
@@ -1,17 +1,15 @@
-import axios, { AxiosRequestConfig, AxiosResponse, Method } from 'axios';
+import axios, { AxiosRequestConfig, AxiosResponse } from 'axios';
 
 import get from './getter';
 import localStorage from './localStorage';
-import { ApiResponse } from './types/api';
+import { ApiResponse, UrlConfigType } from './types/api';
 
 export function callApi(
-  url: string,
-  method: Method = 'GET',
+  urlConfig: UrlConfigType,
   config: AxiosRequestConfig = {}
 ): Promise<ApiResponse<any>> {
   return axios({
-    method,
-    url,
+    ...urlConfig,
     ...config,
   }).then(
     (result: AxiosResponse<any>) => {
@@ -38,20 +36,18 @@ export function callApi(
   );
 }
 
-/*
- * Method to send api request.
- * @param {string} endpoint URL to send request to
- * @param {Method} methodType Request method, default is 'GET'
- * @param {any} data Request payload
- * @param {AxiosRequestConfig} config Axios request config
- * @param {boolean} multipartFormData If request has FormData
- * @param {boolean} withToken For signed requests with token from local storage
- * @returns {ApiResponse} Contains data in field response, if status was 200,
- * otherwise contains error data in fields error and errorData.
+/**
+ * Метод для отправки запроса к api.
+ * @param {UrlConfigType} urlConfig URL, на который нужно отправить запрос, вместе с методом запроса
+ * @param {any} data Тело запроса либо GET-параметры в виде объекта
+ * @param {AxiosRequestConfig} config Конфиг axios
+ * @param {boolean} multipartFormData Содержит ли запрос данные формы
+ * @param {boolean} withToken Для запросов с токеном из local storage
+ * @returns {ApiResponse} Если статус ответа 200, возвращает поле response с ответом от сервера,
+ * иначе поля error and errorData с информацией об ошибке.
  */
 export default function api(
-  endpoint: string,
-  methodType: Method = 'GET',
+  urlConfig: UrlConfigType,
   data: null | undefined | any = {},
   config: AxiosRequestConfig = {},
   multipartFormData = false,
@@ -61,7 +57,7 @@ export default function api(
 
   if (
     (queryConfig.data === null || queryConfig.data === undefined) &&
-    methodType !== 'GET'
+    urlConfig.method !== 'GET'
   ) {
     queryConfig.data = data;
   }
@@ -95,9 +91,7 @@ export default function api(
     });
   }
 
-  const queryMethodType = methodType.toUpperCase() as Method;
-
-  if (queryMethodType === 'GET') {
+  if (urlConfig.method === 'GET') {
     queryConfig.params = {
       ...data,
       uid: localStorage.getItem('userId'),
@@ -108,7 +102,7 @@ export default function api(
     };
   }
 
-  return callApi(endpoint, queryMethodType, queryConfig)
+  return callApi(urlConfig, queryConfig)
     .then((response) => ({ response }))
     .catch((error) => {
       return {